Fix race condition while loading lights
This fixes a race condition where the renderer could start
reading the vector of lights at the same time that newly-loaded
lights were being added to it, thus causing a segfault or other
weird behavior.

This is done by creating separate queues and a corresponding mutex
for mutations to the list of lights in the renderer and flushing
them before starting to draw the lights. Because this is basically
just copying pointers, the actual timespan the mutex is held by the
renderer is very low.